Seperate DOM from index file
